Enhance: Cleanup c/cpp code
Currently, the code style of miluvs-storage is not entirely consistent.
The current changes aim to unify the code style as much as possible and
modify some code that does not follow common C++ practices, including:

- Using get() on shared_ptr/unique_ptr to obtain a raw pointer for method
  calls. Which is unnecessary and dangerous.
- Removed class-private unused variables and local unused variables.
- Implicit type conversions, ex `for (int i = 0; i < size_t)`, which can
  be dangerous.
- Unnecessary `std::move`, such as using `std::move` in cases where Return
  Value Optimization (RVO) already applies.
- Missing `std::move` (resulting in extra memory copies), such as when a
  local variableis assigned and then passed to `push_back` or `emplace_back`.
- Replacing most `push_back` calls with `emplace_back`, as `push_back` always
  incurs an additional memory copy.
- Renaming some methods that used underscore naming to camelCase.

This commit does not yet cover all code comprehensively—only parts that were
encountered during code review have been modified.

+++ cpp/src/common/metadata.cpp
@@ -295,9 +295,9 @@ Result<std::shared_ptr<PackedFileMetadata>> PackedFileMetadata::Make(std::shared
   }
   auto group_fields = GroupFieldIDList::Deserialize(group_field_id_list_meta.ValueOrDie());
   std::map<FieldID, ColumnOffset> field_id_mapping;
-  for (int path = 0; path < group_fields.num_groups(); path++) {
+  for (size_t path = 0; path < group_fields.num_groups(); path++) {
     auto field_ids = group_fields.GetFieldIDList(path);
-    for (int col = 0; col < field_ids.size(); col++) {
+    for (size_t col = 0; col < field_ids.size(); col++) {
       FieldID field_id = field_ids.Get(col);
       field_id_mapping[field_id] = ColumnOffset(path, col);
     }
